web/requirements.txt
google-cloud-bigquery
gunicorn
Flask
requests
